always a request origin, special origin "@" for local
it's now always required to specify what origins (or apps or tags) are allowed
in an Authorization, and all requests have an effective origin. the new special
origin literal specifier "@" matches the local origin. this is to close a
potential security issue where malicious HTML+JS could be uploaded to a
writable directory and then redirected to, where it would then run with
"no/local origin" permission (especially if logged in locally too), which
might bypass the user's intended application permission restrictions.

the implications of this issue are still under investigation. write permission
especially should be restricted to apps by acl:app (and more specific than
origin).